Embed ProviderUtils into MetadataService
